fixed useSearchParams
